requirements.txt
cs50
discord.py
git+https://github.com/ThemeParks/ThemeParks_Python.git
python-dotenv
